Implement general solution for dynamic tabs
