--- blocks/report-lost-found-pet/report-lost-found-pet.css
@@ -1,132 +1,0 @@
-/* TODO Some code from here can be moved to styles.css  */
-.columns.contacts > div {
-	justify-content: center;
-	width: 60%;
-  	margin: 0 auto;
-}
-
-.display-none {
-	display: none;
-}	
-
-.lost-found-pet-selector .out-of-form-radio-wrapper>label {
-	display: block;
-	position: relative;
-	padding: 0.5rem 2rem;
-	border: .1875rem solid var(--button-secondary-color);
-	border-radius: .375rem;
-	cursor: pointer;
-	text-align: center;
-}
-
-.lost-found-pet-selector .out-of-form-radio-wrapper>input[type="radio"]:checked+label {
-	background-color: var(--button-secondary-color);
-	color: var(--background-color);
-}
-
-.lost-found-pet-selector .out-of-form-radio-wrapper>input[type="radio"] {
-	opacity: 0;
-	position: absolute;
-	width: unset;
-	height: unset;
-}
-
-.out-of-form-flex-wrapper {
-	display: flex;
-	justify-content: center;
-	flex-wrap: wrap;
-	gap: 1rem; 
-  }
-
-span.tooltip {
-	display: inline-block;
-	position: absolute;
-	bottom: 1rem;
-	right: .5rem;
-	width: 1.5rem;
-	height: 1.5rem;
-	opacity: 0;
-	content: url("/icons/circle-info-solid.svg");
-}
-
-.complex-form-container form {
-	border: none;
-}
-
-.complex-form-container {
-	  border: .1875rem solid var(--button-secondary-color);
-	  border-radius: .625rem;
-	  margin: 2rem;
-}
-
-.complex-form-container .section.padded{
-	margin: 2rem;
-}
-
-.complex-form-container .section.framed{
-	margin: 2rem;
-	border: .1875rem solid var(--cool-background-color);
-}
-
-  /* Theme for pages with headings colored in white */
-.headings-white h1, .headings-white h2, .headings-white h3, .headings-white h4, .headings-white h5 {
-	color: var(--background-color)
-}
-
-.complex-form-container a.button {
-	display: inline-block;
-}
-
-.complex-form-container .columns- > div > div {
-	background-color: var(--button-secondary-color);
-	text-align: center;
-}
-
-.section .columns img  {
-	min-height: 80px;
-}
-
-.form-flex-container {
-	grid-template-columns: 1fr;
-}
-
-.complex-form-container .columns > div {
-	padding-left: 0;
-	align-items: normal;
-}
-
-@media screen and (width >=768px) {
-	form#report-lost-found-pet {
-		background: rgb(225, 157, 225);
-	}
-	
-	.section .columns img {
-		min-height: 100px;
-	}
-
-	.form-flex-container {
-		display: grid;
-		grid-template-columns: 60% 40%;
-		grid-gap: 10px;
-	  }
-
-	  .flex-right {
-		align-self: start;
-		background-color: green;
-		/* padding: 0; */
-		margin: 0;
-	  }
-
-	  .form-flex-container {
-		display: grid;
-		grid-template-columns: 60% 40%;
-		grid-gap: 10px;
-	  }
-
-	  .flex-right {
-		align-self: start;
-		background-color: green;
-		/* padding: 0; */
-		margin: 0;
-	  }
-}
